+++ codec/src/codec.rs
@@ -10,6 +10,12 @@ pub trait Config: Clone + Send + 'static {}
 // Automatically implement `Config` for matching types.
 impl<T: Clone + Send + 'static> Config for T {}
 
+/// A trait for a `Config` type that is also a `RangeBounds<usize>`.
+pub trait RangeConfig: Config + RangeBounds<usize> {}
+
+// Automatically implement `RangeConfig` for matching types.
+impl<T: Config + RangeBounds<usize>> RangeConfig for T {}
+
 /// Trait for types that can be written (encoded) to a buffer.
 pub trait Write {
     /// Encodes this value by writing to a buffer.
@@ -123,52 +129,13 @@ pub trait EncodeFixed: Write + FixedSize {
 // Automatically implement `EncodeFixed` for types that implement `Write` and `FixedSize`.
 impl<T: Write + FixedSize> EncodeFixed for T {}
 
-/// Extension trait providing an ergonomic read method for types requiring no configuration.
-pub trait ReadExt: Read<()> {
-    /// Reads a value using the default `()` config.
-    fn read(buf: &mut impl Buf) -> Result<Self, Error> {
-        <Self as Read<()>>::read_cfg(buf, ())
-    }
-}
-
-// Automatically implement `ReadExt` for types that implement `Read` with no config.
-impl<T: Read<()>> ReadExt for T {}
-
-/// Extension trait providing ergonomic decode method for types requiring no configuration.
-pub trait DecodeExt: Decode<()> {
-    /// Decodes a value using the default `()` config.
-    fn decode(buf: impl Buf) -> Result<Self, Error> {
-        <Self as Decode<()>>::decode_cfg(buf, ())
-    }
-}
-
-// Automatically implement `DecodeExt` for types that implement `Decode` with no config.
-impl<T: Decode<()>> DecodeExt for T {}
-
-/// Extension trait for types that can read a range of items with a configuration.
-pub trait ReadRangeCfgExt<R: Config + RangeBounds<usize>, Cfg: Config>: Read<(R, Cfg)> {
-    fn read_range_cfg(buf: &mut impl Buf, range: R, cfg: Cfg) -> Result<Self, Error> {
-        Self::read_cfg(buf, (range, cfg))
-    }
-}
-
-// Automatically implement `ReadRangeCfgExt` for types that implement `Read` with a range and config.
-impl<R: Config + RangeBounds<usize>, Cfg: Config, T: Read<(R, Cfg)>> ReadRangeCfgExt<R, Cfg> for T {}
-
-/// Extension trait for types that can read a range of items without configuration.
-pub trait ReadRangeExt<R: Config + RangeBounds<usize>>: ReadRangeCfgExt<R, ()> {
-    fn read_range(buf: &mut impl Buf, range: R) -> Result<Self, Error> {
-        Self::read_range_cfg(buf, range, ())
-    }
-}
-
-// Automatically implement `ReadRangeExt` for types that implement `Read` with a range and no config.
-impl<R: Config + RangeBounds<usize>, T: Read<(R, ())>> ReadRangeExt<R> for T {}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::Error;
+    use crate::{
+        extensions::{DecodeExt, ReadExt},
+        Error,
+    };
     use bytes::Bytes;
 
     #[test]
