+++ codec/src/types/mod.rs
@@ -1,2 +1,4 @@
 pub mod net;
 pub mod primitives;
+pub mod tuple;
+pub mod vec;
