+++ p2p/src/authenticated/types.rs
@@ -1,8 +1,10 @@
 use bitvec::{order::Lsb0, vec};
 use bytes::{Buf, BufMut, Bytes};
-use commonware_codec::{util as CodecUtil, varint, Config, Encode, Error, Read, ReadExt, Write};
+use commonware_codec::{
+    util as CodecUtil, varint, Encode, Error, RangeConfig, Read, ReadExt, Write,
+};
 use commonware_cryptography::Verifier;
-use std::{net::SocketAddr, ops::RangeBounds};
+use std::net::SocketAddr;
 
 #[derive(Clone)]
 pub struct PayloadCodecConfig {
@@ -250,7 +252,7 @@ impl Write for Data {
     }
 }
 
-impl<R: Config + RangeBounds<usize>> Read<R> for Data {
+impl<R: RangeConfig> Read<R> for Data {
     fn read_cfg(buf: &mut impl Buf, range: R) -> Result<Self, Error> {
         let channel = u32::read(buf)?;
         let message = Bytes::read_cfg(buf, range)?;
